Add custom attribute in build.gradle.kts, comment out ServerLib
Added a custom attribute called "myOwnAttribute" with a value of "myOwnValue" to the "archives" configurations block in build.gradle.kts to customize project configuration. In addition, This commit comments out the dependency on ServerLib as it seems to be no longer required. This may indicate a change in the dependency usage pattern within the codebase, improving overall project maintainability.