Refactor and simplify chunk processing code
This commit includes a few different refactoring changes to the chunk processing in the FastAsyncWorldEdit project. It improves the thread handling in the `clearHistory` function, ensuring that the `historyWriteLock` is only unlocked if it was locked on the main thread. It also simplifies some of the conditional logic in the `ChunkHolder` and `FabricGetBlocks` classes. This results in cleaner, easier to read code.